recognizer_task/requirements.txt
langchain
python-dotenv
ipykernel

langchain_community

pypdf
wikipedia

langchain-text-splitters

langchain-openai
chromadb

sentence-transformers
langchain-huggingface 

faiss-cpu
langchain-chroma
langchain-core
langchain-groq

streamlit
arxiv
youtube-transcript-api
pytube
pyperclip
pyautogui
pywin32
keyboard
pyinstaller
pydantic
SpeechRecognition 
pyaudio
spacy
transformers
torchaudio
librosa
pyttsx3
python-Levenshtein
pronouncing
openai-whisper
cassio
tiktoken
cassandra-driver
pinecone
pinecone-client 
pinecone-text 
pinecone-notebooks
neo4j
lamini